requirements.txt
fastapi>=0.127.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.21
sqlalchemy>=2.0.0
pydantic>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
psycopg2-binary>=2.9.0
pyjwt>=2.10.0
python-dotenv>=1.0.0
